+++ js/data-value.js
@@ -1,0 +1,26 @@
+const PHOTO_DESCRIPTION = [
+  'Идеально',
+  'Замечательно',
+  'Вот так вот',
+  'Чё каво?',
+  'Воть',
+];
+
+const COMMENT_TEXT = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const AUTHOR_NAME = [
+  'Артём',
+  'Сергей',
+  'Дмитрий',
+  'Владислав',
+  'Пётр',
+];
+
+export {PHOTO_DESCRIPTION, COMMENT_TEXT, AUTHOR_NAME};
